Removed Automated Script Login
Updated the script so it launches the browser and gives you 30 seconds to login and go to settings, instead of doing it autonomously. This is more efficient as the previous version would have you wait several seconds after each step. Also, this is more secure as your login is no longer kept in plaintext.